Move audio transform class to audio_utils
Code cleanup
